+++ packages/state/jest.config.ts
@@ -4,7 +4,8 @@ import type { Config } from "jest";
 const config: Config = {
   ...baseConfig,
   testTimeout: 15000,
-  setupFilesAfterEnv: ["<rootDir>/src/setupTests.ts"],
+  setupFiles: ["<rootDir>/src/setupTests.ts"],
+  setupFilesAfterEnv: ["<rootDir>/src/setupTestsAfterEnv.ts"],
   rootDir: "./",
   transformIgnorePatterns: ["node_modules/(?!(redux-persist-transform-encrypt)/)"],
 };
